+++ aiida/src/main/java/energy/eddie/aiida/models/permission/dataneed/AiidaLocalDataNeed.java
@@ -8,6 +8,7 @@ import energy.eddie.api.agnostic.aiida.AiidaAsset;
 import energy.eddie.api.agnostic.aiida.AiidaSchema;
 import energy.eddie.api.agnostic.aiida.ObisCode;
 import energy.eddie.api.agnostic.aiida.ObisCodeConverter;
+import energy.eddie.cim.agnostic.PermissionCommand;
 import energy.eddie.dataneeds.needs.aiida.AiidaDataNeed;
 import energy.eddie.dataneeds.needs.aiida.AiidaDataNeedInterface;
 import energy.eddie.dataneeds.utils.cron.CronExpressionConverter;
@@ -62,9 +63,13 @@ public abstract class AiidaLocalDataNeed implements AiidaDataNeedInterface {
     @JsonProperty
     protected boolean acknowledgementRequired;
 
-    @Column(nullable = false, name = "allow_transmission_control")
+    @ElementCollection(fetch = FetchType.EAGER)
+    @CollectionTable(name = "aiida_local_data_need_permission_commands",
+            joinColumns = @JoinColumn(name = "data_need_id"))
+    @Column(name = "permission_command")
+    @Enumerated(EnumType.STRING)
     @JsonProperty
-    protected boolean allowTransmissionControl;
+    protected Set<PermissionCommand.Action> allowedPermissionCommands;
 
     @ElementCollection(fetch = FetchType.EAGER)
     @CollectionTable(name = "aiida_local_data_need_schemas", joinColumns = {@JoinColumn(name = "data_need_id", referencedColumnName = "data_need_id")})
@@ -103,7 +108,7 @@ public abstract class AiidaLocalDataNeed implements AiidaDataNeedInterface {
         this.asset = dataNeed.asset();
         this.dataTags = Objects.requireNonNullElse(dataNeed.dataTags(), Set.of());
         this.acknowledgementRequired = dataNeed.acknowledgementRequired();
-        this.allowTransmissionControl = dataNeed.allowTransmissionControl();
+        this.allowedPermissionCommands = Objects.requireNonNullElse(dataNeed.allowedPermissionCommands(), Set.of());
     }
 
     public String name() {
@@ -141,8 +146,8 @@ public abstract class AiidaLocalDataNeed implements AiidaDataNeedInterface {
     }
 
     @Override
-    public boolean allowTransmissionControl() {
-        return allowTransmissionControl;
+    public Set<PermissionCommand.Action> allowedPermissionCommands() {
+        return allowedPermissionCommands;
     }
 
     @Override