+++ aiida/ui/src/components/PermissionDetails.vue
@@ -282,9 +282,13 @@ onClickOutside(target, () => (showToolTip.value = false))
           </dd>
         </div>
       </div>
-      <div v-if="permission.dataNeed.allowTransmissionControl" class="permission-field">
-        <dt>{{ t('permissions.dropdown.allowTransmissionControl') }}</dt>
-        <dd>{{ permission.dataNeed.allowTransmissionControl }}</dd>
+      <div v-if="permission.dataNeed.allowedPermissionCommands?.length" class="permission-field">
+        <dt>{{ t('permissions.dropdown.allowedPermissionCommands') }}</dt>
+        <div class="column schema">
+          <dd v-for="command in permission.dataNeed.allowedPermissionCommands" :key="command">
+            {{ command }}
+          </dd>
+        </div>
       </div>
       <div v-if="permission.dataNeed.acknowledgementRequired" class="permission-field">
         <dt>{{ t('permissions.dropdown.acknowledgementRequired') }}</dt>