+++ data-needs/src/test/resources/test-data-needs-management-api.sql
@@ -53,7 +53,6 @@ CREATE TABLE IF NOT EXISTS data_needs.outbound_aiida_data_need
     policy_link                text                        NOT NULL,
     purpose                    text                        NOT NULL,
     transmission_schedule      VARCHAR(36)                 NOT NULL,
-    allow_transmission_control boolean                     NOT NULL,
     acknowledgement_required   boolean                     NOT NULL,
     asset                      text                        NOT NULL
 );
@@ -68,7 +67,6 @@ CREATE TABLE IF NOT EXISTS data_needs.inbound_aiida_data_need
     policy_link                text                        NOT NULL,
     purpose                    text                        NOT NULL,
     transmission_schedule      VARCHAR(36)                 NOT NULL,
-    allow_transmission_control boolean                     NOT NULL,
     acknowledgement_required   boolean                     NOT NULL,
     asset                      text                        NOT NULL
 );
@@ -85,6 +83,13 @@ CREATE TABLE IF NOT EXISTS data_needs.aiida_data_need_schemas
     data_need_id uuid NOT NULL,
     schema       text NOT NULL,
     PRIMARY KEY (data_need_id, schema)
+);
+
+CREATE TABLE IF NOT EXISTS data_needs.aiida_data_need_permission_commands
+(
+    data_need_id       uuid NOT NULL,
+    permission_command text NOT NULL,
+    PRIMARY KEY (data_need_id, permission_command)
 );
 
 CREATE TABLE IF NOT EXISTS data_needs.relative_duration